refactor(forks/lstar): rename spec base and fold timeline into fork choice
Rename the internal mixin base module from _contract.py to _base.py and
the class LstarSpecContract to LstarSpecBase, aligning the vocabulary with
its ForkProtocol parent.

Fold the timeline mixin into the fork-choice mixin. on_tick is a fork-choice
store handler, so on_tick and tick_interval now sit beside the interval
actions they drive (accept_new_attestations, update_safe_target, update_head)
rather than living in a separate module.

Drop update_safe_target from the shared base: once timeline merges into fork
choice it is no longer invoked across a mixin boundary, so the base no longer
needs to declare it. This also removes the misleading asymmetry of declaring
update_safe_target while update_head was absent.
