+++ src/lean_spec/spec/forks/lstar/spec.py
@@ -2,7 +2,7 @@
 
 from typing import ClassVar
 
-from lean_spec.spec.forks.lstar._contract import LstarSpecContract, LstarStore
+from lean_spec.spec.forks.lstar._base import LstarSpecBase, LstarStore
 from lean_spec.spec.forks.lstar.aggregation import AggregationMixin
 from lean_spec.spec.forks.lstar.block_production import BlockProductionMixin
 from lean_spec.spec.forks.lstar.containers import (
@@ -18,7 +18,6 @@ from lean_spec.spec.forks.lstar.containers import (
 from lean_spec.spec.forks.lstar.fork_choice import ForkChoiceMixin
 from lean_spec.spec.forks.lstar.signatures import SignatureMixin
 from lean_spec.spec.forks.lstar.state_transition import StateTransitionMixin
-from lean_spec.spec.forks.lstar.timeline import TimelineMixin
 from lean_spec.spec.forks.lstar.validator_duties import ValidatorDutiesMixin
 from lean_spec.spec.forks.protocol import ForkProtocol
 
@@ -31,9 +30,8 @@ class LstarSpec(
     BlockProductionMixin,
     ForkChoiceMixin,
     AggregationMixin,
-    TimelineMixin,
     ValidatorDutiesMixin,
-    LstarSpecContract,
+    LstarSpecBase,
 ):
     """Lstar fork."""
